Fix warning from UnitTestVecFromPortalPermute.cxx
One of the CI jobs (ubuntu2204_gcc12) was warning about this unit test
using a component of the index Vec that may be uninitialized. This
warning was mislead. It was noticing that some of the components of a
`VecVariable` were not used, but those components were never referenced.
Get around this problem by putting the temporary indices in a
`std::vector` and then wrapping the C array in a `VecCConst`.