Remove warning about unused variables
This happened in a plane intersection code that used a subsequent
ray/plane intersection. The plane intersection first checked that
the planes are not parallel. It then uses the ray intersection to
find points on the intersection line. This ray intersection checks
that the ray is not parallel to the plane, but from the previous
check we already know that the ray is not parallel.

Get around the warning that initializes the variables. It
initializes to NaN to verify that the two checks did not disagree.