Enable languages in CMake configuration file
When you link to Viskores from an external CMake package (i.e., by using
`find_package`), it now enables the device languages you might need to compile
the code. Previously, when you loaded the Viskores package and linked your
files, you were likely to get compile problems if Viskores was compiled with
CUDA or HIP.

The problem was that the host code is likely unaware of what if any device is
being compiled. Thus, when it tries to compile Viskores code, it may be using
the wrong compiler.

This fix calls the CMake `enable_language` from the `ViskoresConfigure.cmake`
that is loaded whenever the Viskores package is loaded.