Simplify stream surface filter
The stream surface filter compiled in its own particle tracer. However,
the particle tracing was not any different than the standard stream
tracer. The filter was changed to instead subclass the streamline
filter and replace the cells with a surface. This simplifies the
implementation, speeds up compiles, and reduces the library size.