Simplify code for ArrayHandleSOA portal
The ArrayPortalSOA had some clever template code to set and get values
in multiple arrays. However, we have found that some device compilers
get bogged down resolving all these template parameters in kernels and
the complexity of the code, so the code has been significantly
simplified.

We have also removed support for div and mod in SOAStride. This
reduces the storage and code required for the array portal.

However, since div and mod are not supported there, some tests need
to be cleaned up. Also needed some exceptions suppressed and replaced
with workarounds and warnings.